+++ src/core/t-function.c
@@ -130,6 +130,60 @@ void TO_Function(REBVAL *out, enum Reb_Kind kind, const REBVAL *arg)
 
 
 //
+//  MF_Function: C
+//
+void MF_Function(REB_MOLD *mo, const RELVAL *v, REBOOL form)
+{
+    UNUSED(form);
+
+    Pre_Mold(mo, v);
+
+    Append_Codepoint_Raw(mo->series, '[');
+
+    // !!! The system is no longer keeping the spec of functions, in order
+    // to focus on a generalized "meta info object" service.  MOLD of
+    // functions temporarily uses the word list as a substitute (which
+    // drops types)
+    //
+    REBARR *words_list = List_Func_Words(v, TRUE); // show pure locals
+    Mold_Array_At(mo, words_list, 0, 0);
+    Free_Array(words_list);
+
+    if (IS_FUNCTION_INTERPRETED(v)) {
+        //
+        // MOLD is an example of user-facing code that needs to be complicit
+        // in the "lie" about the effective bodies of the functions made
+        // by the optimized generators FUNC and PROC...
+
+        REBOOL is_fake;
+        REBARR *body = Get_Maybe_Fake_Func_Body(&is_fake, const_KNOWN(v));
+
+        Mold_Array_At(mo, body, 0, 0);
+
+        if (is_fake)
+            Free_Array(body); // was shallow copy
+    }
+    else if (IS_FUNCTION_SPECIALIZER(v)) {
+        //
+        // !!! Interim form of looking at specialized functions... show
+        // the frame
+        //
+        //     >> source first
+        //     first: make function! [[aggregate index] [
+        //         aggregate: $void
+        //         index: 1
+        //     ]]
+        //
+        REBVAL *exemplar = KNOWN(VAL_FUNC_BODY(v));
+        Mold_Value(mo, exemplar);
+    }
+
+    Append_Codepoint_Raw(mo->series, ']');
+    End_Mold(mo);
+}
+
+
+//
 //  REBTYPE: C
 //
 REBTYPE(Function)
